fix custom logic

+++ scripts/misc.sh
@@ -9,10 +9,10 @@ if [[ "${RELEASE_BRANCH%:*}" == "openwrt" ]]; then
     sed -i '/# setup misc settings/ a\mv \/www\/luci-static\/resources\/view\/status\/include\/29_temp.js \/www\/luci-static\/resources\/view\/status\/include\/17_temp.js' files/etc/uci-defaults/99-init-settings.sh
 elif [[ "${RELEASE_BRANCH%:*}" == "immortalwrt" ]]; then
     echo "$RELEASE_BRANCH"
-    if [[ "$(echo "${RELEASE_BRANCH#*:}" | awk -F '.' '{print $1}')" == "23" ]]; then
+    if [[ "$(echo "${RELEASE_BRANCH#*:}" | awk -F '.' '{print $1}')" == "23" && "$PROFILE" == "rpi-4" ]]; then
         cp packages/luci-app-oled_1.0_all.ipk files/root/luci-app-oled_1.0_all.ipk
-        sed -i '/# setup misc settings/ a\opkg install /root/luci-app-oled_1.0_all.ipk --force-reinstall' files/etc/uci-defaults/99-init-settings.sh
         sed -i '/# setup misc settings/ a\rm /root/luci-app-oled_1.0_all.ipk' files/etc/uci-defaults/99-init-settings.sh
+        sed -i '/# setup misc settings/ a\opkg install /root/luci-app-oled_1.0_all.ipk --force-reinstall' files/etc/uci-defaults/99-init-settings.sh
     fi
 fi
